server/src/main/java/br/com/allcool/config/security/AuthenticationService.java
package br.com.allcool.config.security;


import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import br.com.allcool.person.repository.PersonRepository;
import lombok.AllArgsConstructor;

@Service
@AllArgsConstructor
public class AuthenticationService implements UserDetailsService {

	private final PersonRepository personRepository;
	
	@Override
	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
				
		return personRepository.findByEmail(email)
				.orElseThrow(() -> new UsernameNotFoundException("Dados inválidos"));
		
	}

}
